// src/shared/enums.ts
export enum BSONType {
  Double = "double",
  String = "string",
  Object = "object",
  Array = "array",
  BinData = "binData",
  Undefined = "undefined",
  ObjectId = "objectId",
  Bool = "bool",
  Date = "date",
  Null = "null",
  Regex = "regex",
  DBPointer = "dbPointer",
  JavaScript = "javascript",
  Symbol = "symbol",
  JavaScriptWithScope = "javascriptWithScope",
  Int = "int",
  Timestamp = "timestamp",
  Long = "long",
  Decimal = "decimal",
  MinKey = "minKey",
  MaxKey = "maxKey",
}